# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthcare Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  caretakerPhone?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt?: Date;
}